requirements.txt
asn1crypto==1.5.1
beautifulsoup4==4.9.3
cachetools==5.2.0

cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.1
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
google==3.0.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.6.0
google-search-results==2.4.1
googleapis-common-protos==1.56.4
googlesearch-python==1.1.0

gunicorn==20.1.0

httplib2==0.20.4
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.24.2
oauthlib==3.2.2
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.2
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
wincertstore==0.2
